# swagger.yaml
swagger: "2.0"
info:
  title: "Mi APP Full"
  description: "API del proyecto de Delilah Resto"
  version: "2.0.0"
  contact: { email: "santiagonicolasbarreto@gmail.com" }
  license: { name: "Nginx", url: "https://nginx.org/LICENSE" }
host: "localhost:3000"
schemes: [ "http"]
paths:
  /api/register:
    post:
      tags:
        - Users
      description: Registra un nuevo usuario
      parameters:
      - name: name
        type: string
        in: formData
        required: true
        description: Nombre de usuario
      - name: password
        type: string
        in: formData
        required: true
        description: Contraseña del usuario
      - name: username
        type: string
        in: formData
        required: true
        description: username
      - name: email
        type: string
        in: formData
        required: true
        description: Correo electronico del usuario
      - name: phone
        type: number
        in: formData
        required: true
        description: Telefono del usuario
      - name: addressBook
        type: array
        in: formData
        required: true
        description: Direccion de envio del usuario
      responses:
        200: 
          description: Sucess
  /api/login:
    post:
      tags: 
        - "Users"
      description: "Login"
      parameters:
      - name: email
        type: string
        in: formData
        required: true
        description : email
      - name: password
        type: string
        in: formData
        required: true
        description : password
      responses:
        200:
          description: Sucess
  /api/user/me:
    get:
      description: View user profile
      tags: 
        - "Users"
      parameters:
      - name: authorization
        type: string
        in: header
        required: true
        description: token
      responses:
        200:
          description: Sucess

    patch:
      description: Update users
      tags: 
        - "Users"
      parameters:
      - name: authorization
        type: string
        in: header
        required: true
        description: token
      - name: name
        type: string
        in: formData
        required: true
        description: Nombre
      - name: username
        type: string
        in: formData
        required: true
        description: Username
      - name: email
        type: string
        in: formData
        required: true
        description: Email
      - name: phone
        type: number
        in: formData
        required: true
        description: Phone
      - name: addressBook
        type: array
        in: formData
        required: true
        description: ShippingAddress
      responses:
        200:
          description: Sucess
...